yescrypt: extract `pwxform` module; add comments
Moves all code related to smix and the parallel wide tranformation
(pwxform) into the `pwxform` module.
